Move `user_record.py` from `utils` to `models`
