+++ apps/api/src/admin/summary_handler.py
@@ -2,9 +2,9 @@ from collections import Counter
 
 from pydantic import BaseModel, TypeAdapter
 
+from models.user_record import ApplicantStatus, Role
 from services import mongodb_handler
 from services.mongodb_handler import Collection
-from utils.user_record import ApplicantStatus, Role
 
 
 class ApplicantSummaryRecord(BaseModel):
